fix: fix AddText.tsx
